Add mediator for GetStandardOptionKsbs
