+++ src/SFA.DAS.Courses.Api/Controllers/StandardsController.cs
@@ -1,13 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using MediatR;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
 using SFA.DAS.Courses.Api.ApiResponses;
 using SFA.DAS.Courses.Application.Courses.Queries.GetStandard;
+using SFA.DAS.Courses.Application.Courses.Queries.GetStandardOptionKsbs;
 using SFA.DAS.Courses.Application.Courses.Queries.GetStandardsList;
 using SFA.DAS.Courses.Domain.Search;
 
@@ -71,18 +69,13 @@ namespace SFA.DAS.Courses.Api.Controllers
         [Route("{id}/options/{option}/ksbs")]
         public async Task<IActionResult> GetOptionKsbs(string id, string option)
         {
-            return Ok(new GetStandardOptionsResponse
+            var queryResult = await _mediator.Send(new GetStandardOptionKsbsQuery
             {
-                KSBs = new StandardOptionKsb[]
-                {
-                   new StandardOptionKsb
-                   {
-                       Type = KsbType.Knowledge,
-                       Key = "k1",
-                       Detail = "core_knowledge_1",
-                   }
-                }
+                StandardId = id,
+                StandardOption = option,
             });
+
+            return Ok(queryResult);
         }
 
         [HttpGet]
